pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glasstape-policy-builder-mcp"
version = "1.0.0"
description = "Transform natural language into Cerbos policies via MCP"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "GlassTape", email = "hello@glasstape.ai"}
]
keywords = ["mcp", "cerbos", "policy", "ai", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# LLM providers for server-side natural language parsing
anthropic = ["anthropic>=0.18.0"]
bedrock = ["boto3>=1.28.0"]
openai = ["openai>=1.0.0"]

# All LLM providers
llm = [
    "anthropic>=0.18.0",
    "boto3>=1.28.0", 
    "openai>=1.0.0",
]

# Development tools
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
glasstape-policy-builder-mcp = "glasstape_policy_builder.server:main"

[project.urls]
Homepage = "https://glasstape.ai"
Repository = "https://github.com/glasstape/glasstape-policy-builder-mcp"
Issues = "https://github.com/glasstape/glasstape-policy-builder-mcp/issues"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"